% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gptCheckError.R
\name{gptCheckError}
\alias{gptCheckError}
\title{Title}
\usage{
gptCheckError(ErrMessage)
}
\arguments{
\item{ErrMessage}{error message, it is strongly suggested to use raw string
library(tidyverse)
ErrMessage = r"----( paste the material from the output )----"}
}
\description{
Title
}
\examples{
> x=matrix( rnorm(20), nrow=4, ncol=5)
> y=matrix( rnorm(20), nrow=4, ncol=5)
> x\%*\%y
Error in x \%*\% y : non-conformable arguments

Err = r"----( > x=matrix( rnorm(20), nrow=4, ncol=5)
> y=matrix( rnorm(20), nrow=4, ncol=5)
> x\%*\%y
Error in x \%*\% y : non-conformable arguments
)----"

gptCheckError(Err)


}
