% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gptLatexTable.R
\name{gptLatexTable}
\alias{gptLatexTable}
\title{Title}
\usage{
gptLatexTable(nrow, ncol, entries)
}
\arguments{
\item{nrow}{the number of rows}

\item{ncol}{the number of columes}

\item{entries}{the entries of the table to be produced}
}
\description{
this function generates the code for producing a "nrow" x "ncol" table for LaTeX with a given "entries".
}
\examples{
x=matrix(rnorm(12), nrow=3,ncol=4)
LatexTable(3,4, round( t(x), digits=3 )  )
}
