\name{Rgpt-package}
\alias{Rgpt-package}
\alias{Rgpt}
\docType{package}
NA
\title{
\packageTitle{Rgpt}
}
\description{
\packageDescription{Rgpt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Rgpt}
\packageIndices{Rgpt}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Rgpt}

Maintainer: \packageMaintainer{Rgpt}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
