% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgpt.R
\name{Rgpt}
\alias{Rgpt}
\title{Title}
\usage{
Rgpt(inquiry)
}
\arguments{
\item{inquiry}{the question to be sent to chatgpt}
}
\value{

}
\description{
Title
}
\examples{
question="Using R, write a sample code for running linear regression model"
Rgpt(question)



}
