% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgpt_completions.R
\name{Rgpt_completions}
\alias{Rgpt_completions}
\title{Title}
\usage{
Rgpt_completions(
  history,
  prompt,
  openai_api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{history}{the historical record of the conversation with chatgpt}

\item{prompt}{new inquiry}

\item{openai_api_key}{an OPEN_API_KEY is required.}
}
\value{

}
\description{
Title
}
\examples{



}
